+++ tests/gtests/internals/test_cpu_ir.cpp
@@ -1570,663 +1570,4 @@ TEST(IntegrationTests, LoadU8SelectMask) {
         EXPECT_FLOAT_EQ(dst[simd_w + i], -12345.f) << "tail pad lane " << i;
 }
 
-// Arguments for the online-softmax tile epilogue kernel. A tile of `seq_q`
-// score rows of `w` elements each is updated in place; per row, the running
-// max/denominator and the tile's renormalization coefficient are read and
-// written through scalar pointers, matching the per-row state the fused SDPA
-// kernel carries across KV tiles. scores holds seq_q*w floats (row i at i*w);
-// m/l/old_coef hold seq_q floats (row i at i); scale is shared by all rows.
-// cond/fill drive the optional select mask: a masked-out lane takes fill.
-struct softmax_row_args_t {
-    float *scores; // in: raw scores rows; out: normalized probabilities P
-    const float *scale; // scalar softmax scale (shared by all rows)
-    float *m; // in: running row max m_old; out: m_new (one per row)
-    float *l; // in: running denominator l_old; out: l_new (one per row)
-    float *old_coef; // out: corr*l_old/l_new per row (renormalizes running acc)
-    const uint8_t *cond; // select condition bytes (seq_q*w, row i at i*w)
-    const float *fill; // scalar fill for masked-out lanes (shared by all rows)
-};
-
-// Builds the online-softmax epilogue for a tile of `seq_q` score rows, each of
-// width `w` (any w >= 1; the ragged tail beyond the last full simd_w block is
-// handled with masked loads/stores). Mirrors the per-row scalar epilogue in
-// sdp_fused_brgemm.cpp for a single KV tile. With `has_select`, pass 1 also
-// applies the attention select mask: uint8 condition bytes are widened and
-// turned into a lane mask (vload_u8 -> vcmp_ne_zero), then vblend swaps in the
-// broadcast `fill` scalar. The keep sense follows the fused kernel: fusiable
-// keeps the score where cond != 0, non-fusiable where cond == 0 -- encoded by
-// the vblend operand order, no invert op needed.
-// The rows are processed by a runtime loop over seq_q (inlined when seq_q == 1);
-// the score and per-row state pointers advance one row per iteration. Per row
-// the op chain is: scale -> (select) -> running row max -> exp(scaled - m_new)
-// -> running denominator -> divide by l_new. The scalar running state is kept
-// in broadcast vectors so every step is one of the vector ops added for this
-// framework, with no float-immediate op needed (old_coef and the normalization
-// use vdiv). The tail's unused lanes are neutralized before each reduction with
-// `vblend`: seeded with m_old for the max (never beats the running max) and 0
-// for the sum. The first KV tile (m_old == -inf, l_old == 0) needs no explicit
-// guard: the eltwise exp saturates -inf to exactly 0, so corr == 0 and clo ==
-// l_old*corr == 0 fall out, and m_new is just the tile max. The scale, fill and
-// tail mask are loop invariant, set up once; the mask vreg stays live across
-// all rows (masked ops assert it is never spilled).
-ir_t build_softmax_tile_ir(
-        int seq_q, int w, bool has_select = false, bool fusiable = true) {
-    const int n_blk = w / simd_w;
-    const int tail = w % simd_w;
-    const dim_t vbytes = simd_w * (dim_t)sizeof(float);
-    const dim_t tail_off = n_blk * vbytes;
-    const dim_t fsz = (dim_t)sizeof(float);
-
-    ir_t ir;
-
-    // Row pointers: advanced one row per loop iteration.
-    const vreg_t sc_ptr = ir.new_gpr();
-    ir.load_param(sc_ptr, offsetof(softmax_row_args_t, scores));
-    const vreg_t m_ptr = ir.new_gpr();
-    ir.load_param(m_ptr, offsetof(softmax_row_args_t, m));
-    const vreg_t l_ptr = ir.new_gpr();
-    ir.load_param(l_ptr, offsetof(softmax_row_args_t, l));
-    const vreg_t oc_ptr = ir.new_gpr();
-    ir.load_param(oc_ptr, offsetof(softmax_row_args_t, old_coef));
-    // Loop invariant: the scale is shared by every row.
-    const vreg_t scale_ptr = ir.new_gpr();
-    ir.load_param(scale_ptr, offsetof(softmax_row_args_t, scale));
-
-    // Select mask inputs: cond advances one row per iteration (one byte per
-    // score), fill is a loop-invariant scalar shared by every row.
-    vreg_t cond_ptr = vreg_t::none;
-    vreg_t fill_ptr = vreg_t::none;
-    if (has_select) {
-        cond_ptr = ir.new_gpr();
-        ir.load_param(cond_ptr, offsetof(softmax_row_args_t, cond));
-        fill_ptr = ir.new_gpr();
-        ir.load_param(fill_ptr, offsetof(softmax_row_args_t, fill));
-    }
-
-    // One mask, reused by every masked op and every row, active for `tail`.
-    vreg_t mask = vreg_t::none;
-    if (tail) {
-        mask = ir.new_mask();
-        ir.set_mask_imm(mask, tail);
-    }
-
-    // Online-softmax epilogue for the single row at the current pointers.
-    auto row_body = [&]() {
-        // Scratch shared by both horizontal reductions (overwritten each time).
-        const vreg_t ws = ir.new_vec(data_type::f32);
-
-        // Broadcast the scalar inputs so scalar arithmetic reuses vector ops.
-        const vreg_t scale_bc = ir.new_vec(data_type::f32);
-        ir.vload_masked(scale_bc, scale_ptr, 0, vreg_t::none, 1);
-        ir.vbcast(scale_bc, scale_bc);
-
-        vreg_t fill_bc = vreg_t::none;
-        if (has_select) {
-            fill_bc = ir.new_vec(data_type::f32);
-            ir.vload_masked(fill_bc, fill_ptr, 0, vreg_t::none, 1);
-            ir.vbcast(fill_bc, fill_bc);
-        }
-
-        // Apply the select mask to one scaled block of `n` elements at byte
-        // offset `cond_off` into this row's condition bytes, and return the
-        // vreg holding the result (unchanged when there is no select).
-        auto apply_select = [&](vreg_t blk, dim_t cond_off, int n) -> vreg_t {
-            if (!has_select) return blk;
-            const vreg_t cond = ir.new_vec(data_type::s32);
-            ir.vload_u8(cond, cond_ptr, cond_off, n);
-            const vreg_t cmask = ir.new_vec(data_type::s32);
-            ir.vcmp_ne_zero(cmask, cond);
-            if (fusiable) {
-                // Keep the score where cond != 0: cmask ? score : fill.
-                const vreg_t sel = ir.new_vec(data_type::f32);
-                ir.vbcast(sel, fill_bc);
-                ir.vblend(sel, blk, cmask);
-                return sel;
-            }
-            // Keep the score where cond == 0: cmask ? fill : score.
-            ir.vblend(blk, fill_bc, cmask);
-            return blk;
-        };
-
-        const vreg_t m_old = ir.new_vec(data_type::f32);
-        ir.vload_masked(m_old, m_ptr, 0, vreg_t::none, 1);
-        ir.vbcast(m_old, m_old);
-
-        const vreg_t l_old = ir.new_vec(data_type::f32);
-        ir.vload_masked(l_old, l_ptr, 0, vreg_t::none, 1);
-        ir.vbcast(l_old, l_old);
-
-        // Pass 1: scale each block, store it back, and fold it into the running
-        // max (seeded with m_old so the reduction yields m_new directly).
-        const vreg_t rmax = ir.new_vec(data_type::f32);
-        ir.vbcast(rmax, m_old);
-        for (int b = 0; b < n_blk; b++) {
-            vreg_t blk = ir.new_vec(data_type::f32);
-            ir.vload(blk, sc_ptr, b * vbytes);
-            ir.vmul(blk, scale_bc);
-            blk = apply_select(blk, (dim_t)b * simd_w, simd_w);
-            ir.vstore_masked(sc_ptr, b * vbytes, blk, vreg_t::none, simd_w);
-            ir.vmax(rmax, blk);
-        }
-        if (tail) {
-            vreg_t blk = ir.new_vec(data_type::f32);
-            ir.vload_masked(blk, sc_ptr, tail_off, mask, tail);
-            ir.vmul(blk, scale_bc);
-            blk = apply_select(blk, (dim_t)n_blk * simd_w, tail);
-            ir.vstore_masked(sc_ptr, tail_off, blk, mask, tail);
-            // Unused lanes take m_old so they never win the max.
-            const vreg_t tmax = ir.new_vec(data_type::f32);
-            ir.vbcast(tmax, m_old);
-            ir.vblend(tmax, blk, mask);
-            ir.vmax(rmax, tmax);
-        }
-        ir.vhreduce_max(rmax, ws); // rmax lane 0 = m_new
-        const vreg_t m_new = ir.new_vec(data_type::f32);
-        ir.vbcast(m_new, rmax);
-
-        // corr = exp(m_old - m_new): rescales previous tiles' contributions.
-        const vreg_t corr = ir.new_vec(data_type::f32);
-        ir.vbcast(corr, m_old);
-        ir.vsub(corr, m_new);
-        ir.vexp(corr);
-
-        // Pass 2: P_unnorm = exp(scaled - m_new); accumulate the tile denom.
-        const vreg_t rsum = ir.new_vec(data_type::f32);
-        ir.vzero(rsum);
-        for (int b = 0; b < n_blk; b++) {
-            const vreg_t blk = ir.new_vec(data_type::f32);
-            ir.vload(blk, sc_ptr, b * vbytes);
-            ir.vsub(blk, m_new);
-            ir.vexp(blk);
-            ir.vstore_masked(sc_ptr, b * vbytes, blk, vreg_t::none, simd_w);
-            ir.vadd(rsum, blk);
-        }
-        if (tail) {
-            const vreg_t blk = ir.new_vec(data_type::f32);
-            ir.vload_masked(blk, sc_ptr, tail_off, mask, tail);
-            ir.vsub(blk, m_new);
-            ir.vexp(blk);
-            ir.vstore_masked(sc_ptr, tail_off, blk, mask, tail);
-            // Unused lanes take 0 so they add nothing to the denominator.
-            const vreg_t tsum = ir.new_vec(data_type::f32);
-            ir.vzero(tsum);
-            ir.vblend(tsum, blk, mask);
-            ir.vadd(rsum, tsum);
-        }
-        ir.vhreduce(rsum, ws); // rsum lane 0 = tile_sum
-        const vreg_t tile_sum = ir.new_vec(data_type::f32);
-        ir.vbcast(tile_sum, rsum);
-
-        // clo = l_old*corr; l_new = clo + tile_sum; old_coef = clo / l_new.
-        const vreg_t clo = ir.new_vec(data_type::f32);
-        ir.vbcast(clo, l_old);
-        ir.vmul(clo, corr);
-
-        const vreg_t l_new = ir.new_vec(data_type::f32);
-        ir.vbcast(l_new, clo);
-        ir.vadd(l_new, tile_sum);
-
-        const vreg_t old_coef = ir.new_vec(data_type::f32);
-        ir.vbcast(old_coef, clo);
-        ir.vdiv(old_coef, l_new);
-
-        // Pass 3: normalize P by the running denominator.
-        for (int b = 0; b < n_blk; b++) {
-            const vreg_t blk = ir.new_vec(data_type::f32);
-            ir.vload(blk, sc_ptr, b * vbytes);
-            ir.vdiv(blk, l_new);
-            ir.vstore_masked(sc_ptr, b * vbytes, blk, vreg_t::none, simd_w);
-        }
-        if (tail) {
-            const vreg_t blk = ir.new_vec(data_type::f32);
-            ir.vload_masked(blk, sc_ptr, tail_off, mask, tail);
-            ir.vdiv(blk, l_new);
-            ir.vstore_masked(sc_ptr, tail_off, blk, mask, tail);
-        }
-
-        // Write back this row's scalar running state (lane 0).
-        ir.vstore_masked(m_ptr, 0, m_new, vreg_t::none, 1);
-        ir.vstore_masked(l_ptr, 0, l_new, vreg_t::none, 1);
-        ir.vstore_masked(oc_ptr, 0, old_coef, vreg_t::none, 1);
-    };
-
-    // Advance every row pointer to the next row.
-    auto advance_row = [&]() {
-        ir.add_imm(sc_ptr, w * fsz);
-        ir.add_imm(m_ptr, fsz);
-        ir.add_imm(l_ptr, fsz);
-        ir.add_imm(oc_ptr, fsz);
-        if (has_select) ir.add_imm(cond_ptr, w); // one uint8 byte per score
-    };
-
-    emit_loop_imm(ir, seq_q, row_body, advance_row);
-
-    return ir;
-}
-
-// Scalar reference for one online-softmax row update, matching one row of
-// build_softmax_tile_ir (no select mask).
-void ref_softmax_row(std::vector<float> &scores, float scale, float &m,
-        float &l, float &old_coef) {
-    const int w = (int)scores.size();
-    const float neg_inf = -std::numeric_limits<float>::infinity();
-    const float m_old = m, l_old = l;
-    float m_new = m_old;
-    for (int j = 0; j < w; j++) {
-        scores[j] *= scale;
-        m_new = std::max(m_new, scores[j]);
-    }
-    // corr is 0 for the first tile (m_old == -inf), as in the fused kernel.
-    const float corr = m_old == neg_inf ? 0.f : std::exp(m_old - m_new);
-    float tile_sum = 0.f;
-    for (int j = 0; j < w; j++) {
-        const float e = std::exp(scores[j] - m_new);
-        scores[j] = e;
-        tile_sum += e;
-    }
-    const float l_new = l_old * corr + tile_sum;
-    for (int j = 0; j < w; j++)
-        scores[j] /= l_new;
-    m = m_new;
-    l = l_new;
-    old_coef = (l_old * corr) / l_new;
-}
-
-// Scalar reference for one online-softmax row update with the select mask, as
-// in build_softmax_tile_ir(..., has_select=true). A lane is kept when cond != 0
-// (fusiable) or cond == 0 (non-fusiable); otherwise it takes `fill`.
-void ref_softmax_row_masked(std::vector<float> &scores, float scale,
-        const std::vector<uint8_t> &cond, float fill, bool fusiable, float &m,
-        float &l, float &old_coef) {
-    const int w = (int)scores.size();
-    const float neg_inf = -std::numeric_limits<float>::infinity();
-    const float m_old = m, l_old = l;
-    float m_new = m_old;
-    for (int j = 0; j < w; j++) {
-        float v = scores[j] * scale;
-        const bool c = cond[j] != 0;
-        const bool keep = fusiable ? c : !c;
-        if (!keep) v = fill;
-        scores[j] = v;
-        m_new = std::max(m_new, v);
-    }
-    const float corr = m_old == neg_inf ? 0.f : std::exp(m_old - m_new);
-    float tile_sum = 0.f;
-    for (int j = 0; j < w; j++) {
-        const float e = std::exp(scores[j] - m_new);
-        scores[j] = e;
-        tile_sum += e;
-    }
-    const float l_new = l_old * corr + tile_sum;
-    for (int j = 0; j < w; j++)
-        scores[j] /= l_new;
-    m = m_new;
-    l = l_new;
-    old_coef = (l_old * corr) / l_new;
-}
-
-// Validates the online-softmax tile epilogue end to end: the JIT builder's row
-// update must match the scalar reference across several tile widths. This is
-// the first step-3 milestone, proving the whole builder -> alloc -> emit -> run
-// pipeline computes the real softmax math using the ops from steps 1 and 2. The
-// eltwise exp is a polynomial approximation, so denominator-dependent outputs
-// use a relative tolerance; m_new is a plain max and stays exact.
-TEST(IntegrationTests, SoftmaxOnlineTileRow) {
-    SKIP_IF_NO_AVX2();
-
-    // Widths span pure tails (< simd_w), exact multiples, and multiples plus a
-    // ragged tail, so the masked-tail path and its lane neutralization run.
-    for (int w : {1, 5, 7, simd_w, 9, 15, 2 * simd_w, 17, 23, 4 * simd_w - 1,
-                 4 * simd_w}) {
-        ir_kernel_t kernel(build_softmax_tile_ir(1, w));
-        ASSERT_TRUE(kernel.run_ir_pipeline()) << "w=" << w;
-
-        std::vector<float> scores(w), ref(w);
-        for (int j = 0; j < w; j++) {
-            scores[j] = (float)((j * 5) % 13) * 0.5f - 3.f;
-            ref[j] = scores[j];
-        }
-        const float scale = 0.125f;
-        // Finite running state (a later, non-first KV tile).
-        float m = -1.5f, l = 2.0f;
-        float ref_m = m, ref_l = l, ref_oc = 0.f;
-        ref_softmax_row(ref, scale, ref_m, ref_l, ref_oc);
-
-        float oc = -12345.f;
-        softmax_row_args_t args {scores.data(), &scale, &m, &l, &oc};
-        kernel.run(&args);
-
-        EXPECT_NEAR(m, ref_m, 1e-5f * std::abs(ref_m) + 1e-6f) << "w=" << w;
-        EXPECT_NEAR(l, ref_l, 1e-4f * std::abs(ref_l) + 1e-6f) << "w=" << w;
-        EXPECT_NEAR(oc, ref_oc, 1e-4f * std::abs(ref_oc) + 1e-6f) << "w=" << w;
-        for (int j = 0; j < w; j++)
-            EXPECT_NEAR(scores[j], ref[j], 1e-4f * std::abs(ref[j]) + 1e-6f)
-                    << "w=" << w << " j=" << j;
-    }
-}
-
-// Validates the first-tile path (m_old == -inf, l_old == 0) and the running
-// state carried across tiles: one per-row state is threaded through two
-// successive KV tiles, matching two scalar-reference updates. The first tile
-// exercises the -inf seed (the max reduction and exp both saturate correctly so
-// corr == 0); the second tile then consumes the finite state it produced.
-TEST(IntegrationTests, SoftmaxOnlineFirstTile) {
-    SKIP_IF_NO_AVX2();
-
-    const float neg_inf = -std::numeric_limits<float>::infinity();
-    for (int w : {1, 5, 7, simd_w, 9, 15, 2 * simd_w, 17, 23, 4 * simd_w - 1,
-                 4 * simd_w}) {
-        ir_kernel_t kernel(build_softmax_tile_ir(1, w));
-        ASSERT_TRUE(kernel.run_ir_pipeline()) << "w=" << w;
-
-        const float scale = 0.125f;
-        // Two KV tiles of `w` scores, threaded through one running state that
-        // starts at the first-tile seed.
-        float m = neg_inf, l = 0.f, oc = -12345.f;
-        float ref_m = neg_inf, ref_l = 0.f, ref_oc = -12345.f;
-
-        for (int tile = 0; tile < 2; tile++) {
-            std::vector<float> scores(w), ref(w);
-            for (int j = 0; j < w; j++) {
-                scores[j] = (float)((j * 7 + tile * 3) % 13) * 0.5f - 3.f;
-                ref[j] = scores[j];
-            }
-            ref_softmax_row(ref, scale, ref_m, ref_l, ref_oc);
-
-            softmax_row_args_t args {scores.data(), &scale, &m, &l, &oc};
-            kernel.run(&args);
-
-            EXPECT_NEAR(m, ref_m, 1e-5f * std::abs(ref_m) + 1e-6f)
-                    << "w=" << w << " tile=" << tile;
-            EXPECT_NEAR(l, ref_l, 1e-4f * std::abs(ref_l) + 1e-6f)
-                    << "w=" << w << " tile=" << tile;
-            EXPECT_NEAR(oc, ref_oc, 1e-4f * std::abs(ref_oc) + 1e-6f)
-                    << "w=" << w << " tile=" << tile;
-            for (int j = 0; j < w; j++)
-                EXPECT_NEAR(scores[j], ref[j], 1e-4f * std::abs(ref[j]) + 1e-6f)
-                        << "w=" << w << " tile=" << tile << " j=" << j;
-        }
-    }
-}
-
-// Validates the multi-row tile epilogue: one kernel processes seq_q score rows
-// with a runtime loop, advancing the score and per-row state pointers each
-// iteration. Every row carries its own finite running state and distinct data,
-// so a wrong stride would bleed rows into each other. Each row must match an
-// independent scalar-reference update.
-TEST(IntegrationTests, SoftmaxOnlineTileMultiRow) {
-    SKIP_IF_NO_AVX2();
-
-    const float scale = 0.125f;
-    for (int seq_q : {2, 3, 5}) {
-        for (int w : {1, 7, simd_w, 9, 17, 4 * simd_w - 1, 4 * simd_w}) {
-            ir_kernel_t kernel(build_softmax_tile_ir(seq_q, w));
-            ASSERT_TRUE(kernel.run_ir_pipeline())
-                    << "seq_q=" << seq_q << " w=" << w;
-
-            std::vector<float> scores((size_t)seq_q * w);
-            std::vector<float> m(seq_q), l(seq_q), oc(seq_q, -12345.f);
-            std::vector<std::vector<float>> ref(seq_q, std::vector<float>(w));
-            std::vector<float> ref_m(seq_q), ref_l(seq_q), ref_oc(seq_q);
-            for (int i = 0; i < seq_q; i++) {
-                for (int j = 0; j < w; j++) {
-                    const float v
-                            = (float)(((i + 1) * j * 5 + i * 3) % 13) * 0.5f
-                            - 3.f;
-                    scores[(size_t)i * w + j] = v;
-                    ref[i][j] = v;
-                }
-                // Distinct finite per-row running state.
-                m[i] = -1.5f + 0.25f * i;
-                l[i] = 2.0f + 0.5f * i;
-                ref_m[i] = m[i];
-                ref_l[i] = l[i];
-                ref_oc[i] = -12345.f;
-                ref_softmax_row(ref[i], scale, ref_m[i], ref_l[i], ref_oc[i]);
-            }
-
-            softmax_row_args_t args {
-                    scores.data(), &scale, m.data(), l.data(), oc.data()};
-            kernel.run(&args);
-
-            for (int i = 0; i < seq_q; i++) {
-                EXPECT_NEAR(m[i], ref_m[i], 1e-5f * std::abs(ref_m[i]) + 1e-6f)
-                        << "seq_q=" << seq_q << " w=" << w << " i=" << i;
-                EXPECT_NEAR(l[i], ref_l[i], 1e-4f * std::abs(ref_l[i]) + 1e-6f)
-                        << "seq_q=" << seq_q << " w=" << w << " i=" << i;
-                EXPECT_NEAR(
-                        oc[i], ref_oc[i], 1e-4f * std::abs(ref_oc[i]) + 1e-6f)
-                        << "seq_q=" << seq_q << " w=" << w << " i=" << i;
-                for (int j = 0; j < w; j++)
-                    EXPECT_NEAR(scores[(size_t)i * w + j], ref[i][j],
-                            1e-4f * std::abs(ref[i][j]) + 1e-6f)
-                            << "seq_q=" << seq_q << " w=" << w << " i=" << i
-                            << " j=" << j;
-            }
-        }
-    }
-}
-
-// Validates the select mask fused into pass 1 of the softmax tile epilogue:
-// uint8 condition bytes choose between the scaled score and the fill scalar
-// before the running max/denominator update, in both the fusiable (keep where
-// cond != 0) and non-fusiable (keep where cond == 0) senses. Each row runs the
-// full scale -> select -> softmax chain against an independent scalar reference
-// over widths that exercise the ragged tail. The running state is finite (a
-// later KV tile) so even a fully masked row keeps l_new > 0.
-TEST(IntegrationTests, SoftmaxOnlineTileSelect) {
-    SKIP_IF_NO_AVX2();
-
-    const float scale = 0.125f;
-    const float fill = -30.f;
-    for (bool fusiable : {false, true}) {
-        for (int seq_q : {1, 2, 3}) {
-            for (int w : {1, 5, 7, simd_w, 9, 17, 4 * simd_w - 1, 4 * simd_w}) {
-                ir_kernel_t kernel(
-                        build_softmax_tile_ir(seq_q, w, true, fusiable));
-                ASSERT_TRUE(kernel.run_ir_pipeline())
-                        << "fusiable=" << fusiable << " seq_q=" << seq_q
-                        << " w=" << w;
-
-                std::vector<float> scores((size_t)seq_q * w);
-                std::vector<uint8_t> cond((size_t)seq_q * w);
-                std::vector<float> m(seq_q), l(seq_q), oc(seq_q, -12345.f);
-                std::vector<std::vector<float>> ref(
-                        seq_q, std::vector<float>(w));
-                std::vector<std::vector<uint8_t>> refc(
-                        seq_q, std::vector<uint8_t>(w));
-                std::vector<float> ref_m(seq_q), ref_l(seq_q), ref_oc(seq_q);
-                for (int i = 0; i < seq_q; i++) {
-                    for (int j = 0; j < w; j++) {
-                        const float v
-                                = (float)(((i + 1) * j * 5 + i * 3) % 13) * 0.5f
-                                - 3.f;
-                        scores[(size_t)i * w + j] = v;
-                        ref[i][j] = v;
-                        // Mix zeros and nonzero bytes, some > 127.
-                        const uint8_t c = ((i + j) % 2 == 0)
-                                ? 0
-                                : (uint8_t)(50 + (i * 7 + j * 13) % 200);
-                        cond[(size_t)i * w + j] = c;
-                        refc[i][j] = c;
-                    }
-                    // Distinct finite per-row running state (a later KV tile).
-                    m[i] = -1.5f + 0.25f * i;
-                    l[i] = 2.0f + 0.5f * i;
-                    ref_m[i] = m[i];
-                    ref_l[i] = l[i];
-                    ref_oc[i] = -12345.f;
-                    ref_softmax_row_masked(ref[i], scale, refc[i], fill,
-                            fusiable, ref_m[i], ref_l[i], ref_oc[i]);
-                }
-
-                softmax_row_args_t args {scores.data(), &scale, m.data(),
-                        l.data(), oc.data(), cond.data(), &fill};
-                kernel.run(&args);
-
-                for (int i = 0; i < seq_q; i++) {
-                    EXPECT_NEAR(
-                            m[i], ref_m[i], 1e-5f * std::abs(ref_m[i]) + 1e-6f)
-                            << "fusiable=" << fusiable << " seq_q=" << seq_q
-                            << " w=" << w << " i=" << i;
-                    EXPECT_NEAR(
-                            l[i], ref_l[i], 1e-4f * std::abs(ref_l[i]) + 1e-6f)
-                            << "fusiable=" << fusiable << " seq_q=" << seq_q
-                            << " w=" << w << " i=" << i;
-                    EXPECT_NEAR(oc[i], ref_oc[i],
-                            1e-4f * std::abs(ref_oc[i]) + 1e-6f)
-                            << "fusiable=" << fusiable << " seq_q=" << seq_q
-                            << " w=" << w << " i=" << i;
-                    for (int j = 0; j < w; j++)
-                        EXPECT_NEAR(scores[(size_t)i * w + j], ref[i][j],
-                                1e-4f * std::abs(ref[i][j]) + 1e-6f)
-                                << "fusiable=" << fusiable << " seq_q=" << seq_q
-                                << " w=" << w << " i=" << i << " j=" << j;
-                }
-            }
-        }
-    }
-}
-
-// Arguments for the accumulator renormalization epilogue. After the softmax
-// tile update produces old_coef per row, the running output accumulator is
-// rescaled and the new tile's P*V contribution is added in one pass:
-// acc = old_coef*acc + pv, over seq_q rows of hs_v head-size columns. acc and
-// pv hold seq_q*hs floats (row i at i*hs); old_coef holds seq_q floats.
-struct acc_renorm_args_t {
-    float *acc; // in: running output acc; out: renormalized acc
-    const float *pv; // in: this tile's P*V contribution
-    const float *old_coef; // in: per-row renorm coefficient (corr*l_old/l_new)
-};
-
-// Builds the accumulator renormalization for a tile of `seq_q` rows, each of
-// head size `hs` (any hs >= 1; the ragged tail beyond the last full simd_w
-// block uses masked loads/stores). Mirrors the acc rescale in the fused SDPA
-// kernel that follows the softmax epilogue: acc = old_coef*acc + pv. Rows are
-// processed by a runtime loop over seq_q (inlined when seq_q == 1); the acc, pv
-// and old_coef pointers advance one row per iteration. old_coef is broadcast so
-// the per-row scalar reuses the vector ops; no reduction is needed, so the tail
-// needs no lane neutralization (masked ld/st touch only the active columns).
-ir_t build_acc_renorm_ir(int seq_q, int hs) {
-    const int n_blk = hs / simd_w;
-    const int tail = hs % simd_w;
-    const dim_t vbytes = simd_w * (dim_t)sizeof(float);
-    const dim_t tail_off = n_blk * vbytes;
-    const dim_t fsz = (dim_t)sizeof(float);
-
-    ir_t ir;
-
-    // Row pointers: advanced one row per loop iteration.
-    const vreg_t acc_ptr = ir.new_gpr();
-    ir.load_param(acc_ptr, offsetof(acc_renorm_args_t, acc));
-    const vreg_t pv_ptr = ir.new_gpr();
-    ir.load_param(pv_ptr, offsetof(acc_renorm_args_t, pv));
-    const vreg_t oc_ptr = ir.new_gpr();
-    ir.load_param(oc_ptr, offsetof(acc_renorm_args_t, old_coef));
-
-    // One mask, reused by every masked op and every row, active for `tail`.
-    vreg_t mask = vreg_t::none;
-    if (tail) {
-        mask = ir.new_mask();
-        ir.set_mask_imm(mask, tail);
-    }
-
-    // acc = old_coef*acc + pv for the single row at the current pointers.
-    auto row_body = [&]() {
-        const vreg_t oc_bc = ir.new_vec(data_type::f32);
-        ir.vload_masked(oc_bc, oc_ptr, 0, vreg_t::none, 1);
-        ir.vbcast(oc_bc, oc_bc);
-
-        for (int b = 0; b < n_blk; b++) {
-            const vreg_t acc = ir.new_vec(data_type::f32);
-            ir.vload(acc, acc_ptr, b * vbytes);
-            ir.vmul(acc, oc_bc);
-            const vreg_t pv = ir.new_vec(data_type::f32);
-            ir.vload(pv, pv_ptr, b * vbytes);
-            ir.vadd(acc, pv);
-            ir.vstore_masked(acc_ptr, b * vbytes, acc, vreg_t::none, simd_w);
-        }
-        if (tail) {
-            const vreg_t acc = ir.new_vec(data_type::f32);
-            ir.vload_masked(acc, acc_ptr, tail_off, mask, tail);
-            ir.vmul(acc, oc_bc);
-            const vreg_t pv = ir.new_vec(data_type::f32);
-            ir.vload_masked(pv, pv_ptr, tail_off, mask, tail);
-            ir.vadd(acc, pv);
-            ir.vstore_masked(acc_ptr, tail_off, acc, mask, tail);
-        }
-    };
-
-    // Advance every row pointer to the next row.
-    auto advance_row = [&]() {
-        ir.add_imm(acc_ptr, hs * fsz);
-        ir.add_imm(pv_ptr, hs * fsz);
-        ir.add_imm(oc_ptr, fsz);
-    };
-
-    emit_loop_imm(ir, seq_q, row_body, advance_row);
-
-    return ir;
-}
-
-// Scalar reference for one accumulator renormalization row, matching one row of
-// build_acc_renorm_ir.
-void ref_acc_renorm(
-        std::vector<float> &acc, const std::vector<float> &pv, float old_coef) {
-    for (size_t d = 0; d < acc.size(); d++)
-        acc[d] = old_coef * acc[d] + pv[d];
-}
-
-// Validates the accumulator renormalization tile epilogue: one kernel rescales
-// seq_q accumulator rows (acc = old_coef*acc + pv) with a runtime loop,
-// advancing the acc/pv/old_coef pointers each iteration. Every row carries a
-// distinct old_coef and distinct data, so a wrong stride would bleed rows into
-// each other. Each row must match an independent scalar-reference update.
-TEST(IntegrationTests, AccRenormTile) {
-    SKIP_IF_NO_AVX2();
-
-    for (int seq_q : {1, 2, 3, 5}) {
-        // Head sizes span pure tails, exact multiples, and multiples plus a
-        // ragged tail, so the masked-tail path runs.
-        for (int hs : {1, 7, simd_w, 9, 17, 4 * simd_w - 1, 4 * simd_w}) {
-            ir_kernel_t kernel(build_acc_renorm_ir(seq_q, hs));
-            ASSERT_TRUE(kernel.run_ir_pipeline())
-                    << "seq_q=" << seq_q << " hs=" << hs;
-
-            std::vector<float> acc((size_t)seq_q * hs);
-            std::vector<float> pv((size_t)seq_q * hs);
-            std::vector<float> oc(seq_q);
-            std::vector<std::vector<float>> ref(seq_q);
-            for (int i = 0; i < seq_q; i++) {
-                ref[i].resize(hs);
-                for (int j = 0; j < hs; j++) {
-                    const float a
-                            = (float)(((i + 1) * j * 3 + i * 2) % 11) * 0.5f
-                            - 2.f;
-                    const float p
-                            = (float)(((i + 2) * j * 7 + i) % 13) * 0.25f - 1.f;
-                    acc[(size_t)i * hs + j] = a;
-                    pv[(size_t)i * hs + j] = p;
-                    ref[i][j] = a;
-                }
-                // Distinct per-row renorm coefficient.
-                oc[i] = 0.3f + 0.2f * i;
-                std::vector<float> pv_row(pv.begin() + (size_t)i * hs,
-                        pv.begin() + (size_t)(i + 1) * hs);
-                ref_acc_renorm(ref[i], pv_row, oc[i]);
-            }
-
-            acc_renorm_args_t args {acc.data(), pv.data(), oc.data()};
-            kernel.run(&args);
-
-            for (int i = 0; i < seq_q; i++)
-                for (int j = 0; j < hs; j++)
-                    EXPECT_NEAR(acc[(size_t)i * hs + j], ref[i][j],
-                            1e-5f * std::abs(ref[i][j]) + 1e-6f)
-                            << "seq_q=" << seq_q << " hs=" << hs << " i=" << i
-                            << " j=" << j;
-        }
-    }
-}
-
 } // namespace dnnl
